Remove collection view leftovers
