+++ Package.swift
@@ -229,7 +229,6 @@ var streamChatSourcesExcluded: [String] { [
 var streamChatUIFilesExcluded: [String] { [
     "ChatMessageList/TypingIndicatorView_Tests.swift",
     "ChatMessageList/ChatMessageListVC_Tests.swift",
-    "ChatMessageList/ChatMessageListCollectionViewLayout_Tests.swift",
     "ChatMessageList/ChatMessage/ChatMessageErrorIndicator_Tests.swift",
     "ChatMessageList/ChatMessage/ChatMessageBubbleView_Tests.swift",
     "ChatMessageList/ChatMessage/ChatMessageLayoutOptionsResolver_Tests.swift",
