+++ Sources/StreamChatUI/ChatMessageList/ChatThreadVC.swift
@@ -40,14 +40,9 @@ open class _ChatThreadVC<ExtraData: ExtraDataTypes>:
     open lazy var userSuggestionSearchController: _ChatUserSearchController<ExtraData> =
         channelController.client.userSearchController()
 
-    /// Layout used by the collection view.
-    open lazy var messageListLayout: ChatMessageListCollectionViewLayout = components
-        .messageListLayout
-        .init()
-
     /// View used to display the messages
-    open private(set) lazy var tableView: ChatMessageListTableView<ExtraData> = {
-        let tableView = ChatMessageListTableView<ExtraData>().withoutAutoresizingMaskConstraints
+    open private(set) lazy var tableView: _ChatMessageListTableView<ExtraData> = {
+        let tableView = components.messageListTableView.init().withoutAutoresizingMaskConstraints
         tableView.delegate = self
         tableView.dataSource = self
         return tableView
